combining both json dumps() and loads() to store and retrieve data

+++ files/remember_me.py
@@ -3,12 +3,18 @@
 from pathlib import Path
 import json
 
-username=input("what is your name?")
 
 path =Path("username.json")
-contents=json.dumps(username)
-path.write_text(contents)
+if path.exists():
+    contents=path.read_text()
+    username=json.loads(contents)
+    print(f"Welcome back, {username}")
+else:
+    username=input("what is your name?")
+    contents=json.dumps(username)
+    path.write_text(contents)
+    print(f"we will remember you when you return, {username}")
 
-print(f"We will remember you when you come back, {username}")
+
